convert target to long tensors

+++ code/train_classifier.py
@@ -115,7 +115,7 @@ def train():
 
             # squeeze the input, and put on device
             x = x.reshape(x.shape[0], -1).to(device)
-            y = y.reshape(y.shape[0], -1).to(device)
+            y = y.reshape(y.shape[0], -1).long().to(device)
 
             optimizer.zero_grad()
 
@@ -160,7 +160,7 @@ def eval_on_test(nn, loss_function, dl, device):
         losses = []
         for (x, y) in dl:
             x = x.to(device)
-            y = y.to(device)
+            y = y.long().to(device)
 
             test_pred = nn(x).to(device)
